Add support for multiple workers
